[GREEN] Implements the fix.
